requirements.txt
symphony-bdk-python>=2.4.0
Jinja2~=3.0
